Renaming and expose mor API
