+++ src/utils/decode.ts
@@ -1,6 +1,7 @@
 import { isIwaFile, unzip } from './unzip';
-import { unchunk } from './unchunk';
+import { dechunk } from './dechunk';
 import { IwaObject, splitObjects } from './split';
+import { uncompress } from './uncompress';
 
 export interface ArchiveEntryBase {
   type: 'file' | 'iwa';
@@ -27,8 +28,9 @@ export async function* decode(data: Uint8Array): AsyncIterableIterator<ArchiveEn
   for await(const entry of unzip(data)) {
     if(isIwaFile(entry.name)) {
       const chunks: IwaChunk[] = [];
-      for await(const chunk of unchunk(entry.data)) {
+      for await(const snappyChunk of dechunk(entry.data)) {
         const objects: IwaObject[] = [];
+        const chunk = await uncompress(snappyChunk);
         for await(const message of splitObjects(chunk)) {
           objects.push(message);
         }
