+++ src/utils/index.ts
@@ -1,7 +1,8 @@
 export * from './constants';
+export * from './dechunk';
 export * from './decode';
 export * from './reader';
 export * from './split';
-export * from './unchunk';
+export * from './uncompress';
 export * from './unzip';
 export * from './utils';
