+++ src/utils/utils.ts
@@ -2,6 +2,6 @@ function serializeBigInt(_key: string, value: any) {
   return typeof value === 'bigint' ? value.toString() : value;
 }
 
-export function serialize(message: unknown): string {
+export function asJson(message: unknown): string {
   return JSON.stringify(message, serializeBigInt, 2);
 }
